Add some FWT discussion notes.  Refs #11065.
In discussion with the FWT, we have some suggestions regarding the
installation/activation/enabling story.  If an admin activates the
add-on, that seems a pretty clear indication that they want caching,
so maybe "enable caching" should be on by default.  Also, though it's
covered in the documentation, it's not very clear from the control
panel UI that one needs to do more than just check "enable caching" to
get caching.  So it would be good to have something on that view that
points the user to the "import settings" tab.  Another suggestion that
would be great to see would be to display a portal status message when
"enable caching" is checked but no caching profile has been imported,
like we do with the mail host settings.  These two changes may reduce
some confusion for new users.

svn path=/buildouts/plone-coredev/branches/4.1/; revision=46194