Add a review note about reusing plone.app.querystring.  Refs #9352.
It occurrs to me that this PLIP could have and probably should have
re-used the plone.app.querystring package created for
plone.app.collection which provides their AJAX search query
building.  The stuff used by plone.app.collection is pretty slick
and we should avoid multiple ways of doing things.  It seems like
this ship has already sailed so it may be too late but I wanted to
note it.

svn path=/buildouts/plone-coredev/branches/4.1/; revision=46338