Use plone.app.blob version 1.5.10.
After an unrelated change in plone.app.contenttypes, there is a test
failure on Jenkins in Plone 4.3, Python 2.7, something with an
unexpected 'created' index in the testCatalogAfterImageMigration test
of plone.app.blob.  I suspect this may be randomly failing.
Let's upgrade the minor version to see if that fixes it for the moment.