Pin more packages.
Not sure if 'Build tools' is the correct header for these packages.

Note that FormEncode and MarkupSafe were reported unpinned as mixed
case, but they have to be pinned with lowercase.  With zc.buildout 2
(or 1 with buildout-versions extension), the MixedCase pin would have
worked.  Lowercase apparently works in all cases.