[fc] Repository: Products.CMFPlone Branch: refs/heads/master Date: 2015-08-14T01:46:21+02:00 Author: Maurits van Rees (mauritsvanrees) <maurits@vanrees.org> Commit: https://github.com/plone/Products.CMFPlone/commit/2389666a30d38d064ef6da79623bdf07e67ffc32
folder_listing: avoid rebatching an empty foldercontents batch.

If 'folderContents' was a Batch but it was empty, we would batch this batch,
which works, but gives a DeprecationWarning because in plone.batching
you then start asking for len(self._sequence), which is deprecated on a PloneBatch.

This uses newer style 'A if condition else B' instead of 'if condition and A or B'.

Files changed:
M Products/CMFPlone/skins/plone_content/folder_listing.pt