[fc] Repository: plone.app.contenttypes Branch: refs/heads/master Date: 2015-09-12T03:04:04+02:00 Author: Maurits van Rees (mauritsvanrees) <maurits@vanrees.org> Commit: https://github.com/plone/plone.app.contenttypes/commit/74522d373eac040d9b1d04febf6c4b93b97291ae
Use setuptools 18.2 in bootstrap.

We are installing this version with pip and pinning it in the buildout,
so we should use it while bootstrapping as well.
Otherwise when running 'bin/buildout -N' this fails with a VersionConflict:
it has 18.3.1 in bin/buildout and and the script is called in non-newest mode,
so it cannot update itself.
Alternatively, simply run bin/buildout without '-N',
seems fine.
Again, alternatively, remove the setuptools pinning. 18.3 had a problem, but 18.3.1 is fine again.

Files changed:
M .travis.yml