Commented out five.formlib source.
We are using version 1.0.4. That is over four years old. Since then,
the master branch has introduced an incompatible change,
by running:

  processInputs(self.request, [HTTPRequest.default_encoding])

When starting Zope, the standard zope.conf from
plone.recipe.zope2instance will make this encoding utf-8.  But in
tests, it is still the default of iso-8859-15.  This causes a few test
failures in plone.app.users, plone.app.blob and
plone.app.controlpanel.

Possibly we could override the default_encoding in
Products.PloneTestCase and/or plone.app.testing.

But since five.formlib master has not been touched in three years and
1.0.4 is still the latest release and Plone 5 is not using it anymore,
we should probably just ignore the source and be happy with the
latest release.