[fc] Repository: plone.app.contenttypes Branch: refs/heads/master Date: 2015-11-20T23:46:11-06:00 Author: vangheem (vangheem) <vangheem@gmail.com> Commit: https://github.com/plone/plone.app.contenttypes/commit/9b20dc959d73bdc31d400b9893208333c82ce717
avoid AttributeError here. This is hard to pin point and harder to test. I am getting an error on a production site that needs this fix. It is weird and not reproducible in dev mode though. This is just more safe code so I am contributing this just to prevent others from having the same headache. Sorry for no test, I do not know how to make it break to be honest...

Files changed:
M CHANGES.rst
M plone/app/contenttypes/browser/folder.py