[fc] Repository: plone.portlet.collection Branch: refs/heads/2.1.x Date: 2015-12-07T11:22:11+01:00 Author: Maurits van Rees (mauritsvanrees) <maurits@vanrees.org> Commit: https://github.com/plone/plone.portlet.collection/commit/0ed9c53d907c2f1c3ccaf29e451757d8eaed71fa
Fixed sometimes failing test due to sorting.

Sorting on 'created' could give wrong results,
because the four items of this test are probably created in the same second,
which means the order might be wrong.
On Jenkins number 3 and 0 were in the list, instead of the expected 0 and 1.

Files changed:
M CHANGES.txt
M plone/portlet/collection/tests/test_portlet_collection.py