[fc] Repository: Products.PlonePAS Branch: refs/heads/master Date: 2016-02-12T21:44:20+01:00 Author: Maurits van Rees (mauritsvanrees) <maurits@vanrees.org> Commit: https://github.com/plone/Products.PlonePAS/commit/f241b8c1943de23c3f780de950ce9321e4bbeca9
Fixed AttributeError with Python 2.6 when reading setup.py.

install_requires inadvertently was a list and after this a comma,
which turned it into a tuple with a list as only value.
Strangely, this worked fine in Python 2.7.
But in Python 2.6 we then tried to append to this tuple, which failed.

Files changed:
M CHANGES.rst
M setup.py