Added more version pins.
argparse is required by plone.app.robotframework.
alabaster and snowballstemmer are required by sphinx.
CacheControl is required by plone.versioncheck.
lockfile is required by plone.versioncheck.

A note about the 'persistent' package':

We are also pulling in 'persistent' and not pinning it.
This only ends up in bin/test and a few other scripts, but not in
bin/instance.
We use the persistent library that is shipped inside of the ZODB3 package.

persistent is required by for example the z3c.form tests.
versions.cfg still cautions against adding a pin for the persistent
package:

  Persistent is not compatible with ZODB 3.10 which we are using
  If you have a dependency somewher on persistent, you must
  try to get rid of it.