Downgrade lazr.restfulclient from 0.13.3 to 0.13.1.
That one actually is on PyPI.
Details on this problem are here:
https://bugs.launchpad.net/lazr.restfulclient/+bug/1500460
Main thing is that 0.13.1 is the most recent one that is listed here:
https://pypi.python.org/simple/lazr.restfulclient/

I think PyPI policy has recently changed so that other urls are no
longer offered, so the download url and homepage are no longer given.
That may also be a change in recent setuptools, so maybe with an
earlier setuptools you can still get it.

I have the 0.13.3 file locally and have uploaded it here:
http://pypi.zestsoftware.nl/public/
You can use that in the find-links if you want.