[fc] Repository: Products.CMFPlone Branch: refs/heads/5.0.x Date: 2016-04-20T08:01:47-05:00 Author: Maurits van Rees (mauritsvanrees) <maurits@vanrees.org> Commit: https://github.com/plone/Products.CMFPlone/commit/330df732cd68750e2bda9ff040f22634ab7dcb71
Fixed displaying the body text of a feed item.

This is when render_body is switched on in the Syndication settings.

We were overwriting the definition of the `view` variable, which made
it fail:
AttributeError: 'str' object has no attribute 'context'

Strangely, this does not fail in the tests.

Note that we caught the error so no error was visible to a visitor.

Also note that getting the content-core macro may fail anyway, for
example for a Folder, because its default view does not have this
macro.

Files changed:
M CHANGES.rst
M Products/CMFPlone/browser/syndication/templates/content_core.pt