[fc] Repository: plone.app.jquerytools Branch: refs/heads/master Date: 2016-08-08T15:17:58+02:00 Author: Vincent Fretin (vincentfretin) <vincent.fretin@gmail.com> Commit: https://github.com/plone/plone.app.jquerytools/commit/af45786fd323b7682b06ee4b3f08ae74551739fd
The prepOverlay function can be used on a form instead of a link.
When a button of this form is clicked, the form action attribute is used to
retrieve the overlay. (This was already the case in previous versions,
What's new is the following.) If method="POST", the overlay will be fetched
via a POST with the form payload instead of GET.
The use of a POST instead of GET may be useful to avoid the url length limit.

Files changed:
M CHANGES.rst
M README.rst
M plone/app/jquerytools/browser/overlayhelpers.js
M setup.py