Moved coverage 3.7.1 pin from tests.cfg to versions.cfg.
This overrides the ztk version 3.5.2, which is really old,
and incompatible with current createcoverage.
I think since one or two years this has already bitten add-on developers
who get Travis failures for this, and need to pin either an older
createcoverage, or a newer coverage.

BTW, newest coverage is 4.2, but let's start with simply moving the version pin.