[fc] Repository: plone.app.caching Branch: refs/heads/master Date: 2016-10-04T17:41:22+02:00 Author: Maurits van Rees (mauritsvanrees) <maurits@vanrees.org> Commit: https://github.com/plone/plone.app.caching/commit/ecf2150113b72447488e65d296ae02afe97d2c8a
Set browser.handleErrors to False in one test.

This is when requesting an image scale.
On the plone.namedfile branch for plip 1483 I am getting a KeyError 'x-cache-rule'
when checking the browser headers,
where it would be handier to get an Unauthorized because the user is not allowed to see this image.
(The user probably *should* be allowed to see this image,
but that is something to fix in that plip branch.)

Files changed:
M plone/app/caching/tests/test_profile_with_caching_proxy.py