Add note regarding packages that run fully isolated
plone.recipe.alltests gets a list of ALL packages being used by Plone,
then starts running packages by groups and removing them from that global list.

Finally it runs the packages that are still on that global list.

This note is about all the packages that are not on any group for whichever
reason, either isolation problems or someone forgot to add it to the releavant group.

[ci skip]