[fc] Repository: plone.app.multilingual Branch: refs/heads/master Date: 2016-11-01T10:35:48+01:00 Author: Maurits van Rees (mauritsvanrees) <maurits@vanrees.org> Commit: https://github.com/plone/plone.app.multilingual/commit/522b5cbadcc991e07abb3400f0f450b7311d0cd9
Robot tests: first wait until page contains element.

Only after this, wait until element is visible.
Otherwise you get an error:
Message: Element not found in the cache - perhaps the page has changed since it was looked up
If the element was already on the page, this is of course not needed.
But here the element was added dynamically, so we need this extra step.
At least for Firefox.

Files changed:
M src/plone/app/multilingual/tests/robot/test_add_translation.robot