Update AccessControl to 3.0.12, released December 2015
This then requires an update of ExtensionClass and Acquisition to much newer versions.
So this may be tricky.  We had some problems getting this on Plone 5.

Need newer five.intid, as some of its tests fail.  The newer one fixes something with circular acquisitions.

plone.dexterity fails too, so we need a newer one there, which may be too much.