[fc] Repository: Products.CMFQuickInstallerTool Branch: refs/heads/master Date: 2017-02-02T18:57:19+01:00 Author: Maurits van Rees (mauritsvanrees) <maurits@vanrees.org> Commit: https://github.com/plone/Products.CMFQuickInstallerTool/commit/4a4f6b6a42011e614b7f585bf5d740cdff012b1a
Fixed test that got redirected wrongly.

This went wrong on the Zope4 PLIP.
It looks like self.browser.post changed to not only post,
but also handle any redirects that are done afterwards.
This meant that portal_quickinstaller/installProducts got loaded as GET, which is not allowed.
So we do a GET ourselves first, so a referer is set that is then redirected to.

Files changed:
M Products/CMFQuickInstallerTool/tests/test_browser.py