Revert p.a.linkintegrity to 1.5.8 and check it out.
My 4.3 coredev buildout with P{ython 2.6 said for the 1.5.9 that I just released:

Getting distribution for 'plone.app.linkintegrity==1.5.9'.
While:
  Installing instance.
  Getting distribution for 'plone.app.linkintegrity==1.5.9'.
Error: Couldn't find a distribution for 'plone.app.linkintegrity==1.5.9'.

So I made a release as 1.5.10, to see if that would fix it, but it did not help.

So back to 1.5.8 and a checkout for the moment.