Revert setuptools to 33.1.1
Anything prior to setuptools 33.1.1 should be fine.

Plone uses basic features of setuptools and as our Jenkins CI environment
is already pining 26.1.1 is easier to stick to this version.

As soon as PLIP 2001 https://github.com/plone/Products.CMFPlone/issues/2001 is approved and merged,
we can happily move to virtualenv which solves the problem completely.