zc.buildout 2.9.5 (and setuptools 33.1.1).
This accesses PyPI via https, which is good because PyPI does not listen to http anymore.

Case in point: when you have a previous zc.buildout installed,
and you update the version pin to 2.9.5 and run bin/buildout,
it fails because it cannot find this new version.  Or any new package.
Do a bootstrap, and it will work again.

Workaround would be to add this line with https in the [buildout] section
index = https://pypi.python.org/simple/