Pin collective.MockMailHost = 2.0.0.
My commit f226fbf7eac9fad86e5f975ad03fa5a27e2b775b form yesterday tried to do that,
with some automated help from plone.releaser,
but only the checkout was removed: the pin stayed the same.
Reason is that collective.mockmailhost in lowercase was pinned,
instead of the collective.MockMailHost mixedcase that is the official package name.
So let's use the official mixedcase name.

I wonder if this was the reason for this test failure that started happening on Jenkins 5.2 on Python 3.6 and 3.7 after that commit:
https://jenkins.plone.org/job/plone-5.2-python-3.6/142/testReport/junit/Products.CMFPlone.tests.testQueryCatalog/TestSearchForms/testRenderSearchForm/
The SearchControlpanel from that exception comes from plone.restapi, which is using collective.MockMailHost.
So maybe something goes wrong in test layers.