Revert "On Python 2.7, pin keyring back to 18.0.0."
This reverts commit 75650d0ad7b1a79f635d0c910e03dd2b7bc154bf.

It gives a buildout error on Jenkins, but not locally for me.

See https://jenkins.plone.org/job/plone-5.2-python-2.7/273/console
The error is:

```
Installing releaser.
Version and requirements information containing secretstorage:
  [versions] constraint on secretstorage: 3.1.1
  Requirement of keyring: secretstorage<3
While:
  Installing releaser.
Error: The requirement ('secretstorage<3') is not allowed by your [versions] constraint (3.1.1)
```

But that is strange, because secretstorage is not pinned at all in our versions.
It does not show up in `bin/buildout annotate`.
Locally it does not end up in `bin/fullrelease`.