Remove Products.CMFFormController / portal_form_controller.
Remove the package from the buildout files.
Use branches of CMFPlone and plone.app.upgrade.

We could also remove a CMFFormController line from plone.app.testing, but it's not in the way.
Could be done once we make a plone.app.testing branch for Plone 6 only.

In plone.app.upgrade I remove the portal_form_controller tool.
We could also conditionally remove it: check if the tool is an instance of `ZODB.broken.Broken`.
This is to support the theoretical case that someone explicitly adds the CMFFormController package,
because it is still needed in a specific site.
But same could be done for portal_quickinstaller, and we don't do it there either.