Remove several version pins, move non-core ones to versions-extra.cfg.
versions.cfg should now only have pins for packages used when creating a Plone instance and `bin/test`.
I have split them into four groups to have a bit of structure.
The four groups are based on the GitHub organisations that the packages belong to:

1. Plone
2. Collective
3. Zope, ZODB, z3c, repoze
4. other

Fine to revise if this becomes hard to understand and maintain.