Take over changes from release/5.2-dev branch.
I have released and tagged Plone 5.2.9 on that branch.

This includes one real change in the versions:

Use Sphinx 1.8.5 in all Python versions, instead of 4.5 in Python 3.
This avoids a version conflict with importlib-metadata.
Seen because 'tox -c release/tox.ini -e downloads3' failed with that version conflict:
the new Sphinx requires importlib-metadata 4.4+ and we pin 0.23.

Maybe updating importlib-metadata would have been fine, but not at the last moment.
That could be tested separately.