Add unofficial newer-versions.cfg to 5.2.
This file contains unofficial newer version pins for a few packages.
These are known to work on Plone 5.2, but there are reasons to not include
them officially.  The changes in there may be slightly more adventurous
than is deemed wise at this late stage in the Plone 5.2 release cycle:
currently (November 2023) Plone 5.2 is even out of maintenance support.
Some do not work on Python 2, and we do not want to have two different
version pins for core packages, to avoid confusion.