Upgrade Pillow to 10.2.0.
This should work, now that https://github.com/plone/plone.scale/pull/90 is merged.

I am inclined to say we only do this for Plone 6.1, not 6.0.  The release has backwards incompatibilities.  Should not be noticeable under normal use, but you never know what projects have as extra code.
I am not aware of anything new in Pillow 10 that we would really want to actively use in Plone.  So 9.5 should still be fine for Plone 6.0.