Pin both WebTest and webtest.
In `versions.cfg` we pinned `webtest = 3.0.3`, which was meant as an override for `WebTest = 3.0.1` in Zope 5.12 `versions-prod.cfg`.
I think the spelling may have changed once or twice.

But now when running `tox -e test` on a package that uses http://dist.plone.org/release/6.0-dev/constraints.txt (which I updated earlier today, and which is derived from the versions.cfg), I get an error:

```
The conflict is caused by:
    zope-testbrowser 7.0 depends on WebTest>=2.0.30
    The user requested (constraint) webtest==3.0.1,==3.0.3
```