Do not use plone.cachepurging '[test]' extra.
It is empty, so we can leave it out.
And on Python 3.13 on gh-actions Ubuntu it unexpectedly gives an [error](https://github.com/plone/buildout.coredev/actions/runs/15686144971/job/44189715178?pr=1020):

```
While:
  Installing test.
Error: Couldn't find the required extra.
This means the requirement is incorrect.
If the requirement is itself from software you requested,
then there might be a bug in requested software.
You can ignore this by using 'allow-unknown-extras=true',
however that may simply cause needed software to be omitted.

plone.cachepurging 3.0.2 does not provide the extra 'test'
```

Locally on Mac it is fine.