Revert "borg.localrole 3.1.12"
Go back to 3.1.9.
This reverts commit a24209c0eed6a477c63b96eabb6c64a54fb94d3e.

During release I got an error uploading a file with a name like `borg.localrole.whl/tgz`.  It is no longer accepted by PyPI.  So I used a newer setuptools version, which generates a name with an underscore instead of a dot.  But now that cannot be installed... So go back to the previous version.
And no, that is not 3.1.10 or 3.1.11 because they had incompatible changes.
3.1.9 and 3.1.12 are essentially the same release.